refactor: Change project name to fix name collision on pypi.org
