+++ .github/workflows/linux.yml
@@ -1,4 +1,4 @@
-name: Dailylog linux
+name: Dailylog-linux
 # Run this job on pushes to `main`, and for pull requests. If you don't specify
 # `branches: [main], then this actions runs _twice_ on pull requests, which is
 # annoying.
