+++ dailylog1/__init__.py
