+++ dailylog1/exceptions.py
