+++ dailylog1/version.py
