+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "dailylog"
+name = "dailylog1"
 version = "0.1.2"
 description = ""
 authors = ["Quien Sabe <qs5779@mail.com>"]
@@ -22,8 +22,8 @@ classifiers = [
 ]
 
 packages = [
-    { include = "dailylog" },
-    { include = "dailylog/py.typed" },
+    { include = "dailylog1" },
+    { include = "dailylog1/py.typed" },
 ]
 
 [tool.poetry.scripts]
