+++ setup.cfg
@@ -1,4 +1,4 @@
-# setup.cfg for project dailylog
+# setup.cfg for project dailylog1
 # All configuration for plugins and other utils is defined here.
 # Read more about `setup.cfg`:
 # https://docs.python.org/3/distutils/configfile.html
@@ -51,8 +51,8 @@ exclude =
 # WPS442 => Found outer scope names shadowing
 per-file-ignores =
   tests/*.py: S101 WPS442, E800, S108
-  dailylog/cli.py: WPS216, WPS211
-  dailylog/cache.py: DAR101, DAR102
+  dailylog1/cli.py: WPS216, WPS211
+  dailylog1/cache.py: DAR101, DAR102
   docs/conf.py: NIP323
 
 [isort]
@@ -74,7 +74,7 @@ addopts =
   --strict-config
   --tb=short
   --doctest-modules
-  --cov=dailylog
+  --cov=dailylog1
   --cov-report=term-missing:skip-covered
   --cov-report=html
   --cov-report=xml
